edit filterMap, move get list code
