+++ ncloud/data_source_ncloud_route_tables.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/NaverCloudPlatform/ncloud-sdk-go-v2/ncloud"
-	"github.com/NaverCloudPlatform/ncloud-sdk-go-v2/services/vpc"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
 
@@ -51,68 +49,4 @@ func dataSourceNcloudRouteTablesRead(d *schema.ResourceData, meta interface{}) e
 	}
 
 	return nil
-}
-
-func getRouteTableList(d *schema.ResourceData, config *ProviderConfig) (*vpc.GetRouteTableListResponse, error) {
-	reqParams := &vpc.GetRouteTableListRequest{
-		RegionCode: &config.RegionCode,
-	}
-
-	if v, ok := d.GetOk("vpc_no"); ok {
-		reqParams.VpcNo = ncloud.String(v.(string))
-	}
-
-	if v, ok := d.GetOk("name"); ok {
-		reqParams.RouteTableName = ncloud.String(v.(string))
-	}
-
-	if v, ok := d.GetOk("supported_subnet_type"); ok {
-		reqParams.SupportedSubnetTypeCode = ncloud.String(v.(string))
-	}
-
-	if v, ok := d.GetOk("route_table_no"); ok {
-		reqParams.RouteTableNoList = []*string{ncloud.String(v.(string))}
-	}
-
-	logCommonRequest("data_source_ncloud_route_tables > GetRouteTableList", reqParams)
-	resp, err := config.Client.vpc.V2Api.GetRouteTableList(reqParams)
-
-	if err != nil {
-		logErrorResponse("data_source_ncloud_route_tables > GetRouteTableList", err, reqParams)
-		return nil, err
-	}
-
-	logResponse("data_source_ncloud_route_tables > GetRouteTableList", resp)
-	return resp, nil
-}
-
-func getRouteTableListFiltered(d *schema.ResourceData, config *ProviderConfig) ([]map[string]interface{}, error) {
-	resp, err := getRouteTableList(d, config)
-
-	if err != nil {
-		return nil, err
-	}
-
-	resources := []map[string]interface{}{}
-
-	for _, r := range resp.RouteTableList {
-		instance := map[string]interface{}{
-			"id":                    *r.RouteTableNo,
-			"route_table_no":        *r.RouteTableNo,
-			"name":                  *r.RouteTableName,
-			"description":           *r.RouteTableDescription,
-			"status":                *r.RouteTableStatus.Code,
-			"vpc_no":                *r.VpcNo,
-			"supported_subnet_type": *r.SupportedSubnetType.Code,
-			"is_default":            *r.IsDefault,
-		}
-
-		resources = append(resources, instance)
-	}
-
-	if f, ok := d.GetOk("filter"); ok {
-		resources = ApplyFilters(f.(*schema.Set), resources, resourceNcloudRouteTable().Schema)
-	}
-
-	return resources, nil
 }
